requirements.txt
openai
streamlit
python-dotenv
pypdf2
docx2txt
requests
pandas
langchain
pillow
faiss-cpu
numpy
matplotlib
seaborn
plotly
scikit-learn
sqlalchemy
werkzeug
langchain-community
langchain_openai